separate admin routes from student routes

+++ routes/web.php
@@ -32,15 +32,15 @@ Route::get('/home', 'HomeController@index')->name('home');
 
 Route::get('/admin', function () {
     return view('admin.index');
-});
+})->middleware(['role:admin']);
 
 //============== Courses ====================
 
 Route::group(['middleware' => 'auth'], function () {
     Route::get('/courses', 'CourseController@index')->name('courses.index');
-    Route::get('/courses/create', 'CourseController@create')->name('courses.create');
-    Route::post('/courses', 'CourseController@store')->name('courses.store');
-    Route::get('/courses/{course}/edit', 'CourseController@edit')->name('courses.edit');
-    Route::put('/courses/{course}', 'CourseController@update')->name('courses.update');
-    Route::delete('/courses/{course}', 'CourseController@destroy')->name('courses.destroy');
+    Route::get('/courses/create', 'CourseController@create')->name('courses.create')->middleware(['role:admin']);
+    Route::post('/courses', 'CourseController@store')->name('courses.store')->middleware(['role:admin']);
+    Route::get('/courses/{course}/edit', 'CourseController@edit')->name('courses.edit')->middleware(['role:admin']);
+    Route::put('/courses/{course}', 'CourseController@update')->name('courses.update')->middleware(['role:admin']);
+    Route::delete('/courses/{course}', 'CourseController@destroy')->name('courses.destroy')->middleware(['role:admin']);
 });
